extract dataclasses for agri2030
